common: fix GCodeAnalysisInformation field type

+++ common.go
@@ -305,7 +305,7 @@ type Reference struct {
 // GCodeAnalysisInformation Information from the analysis of the GCODE file.
 type GCodeAnalysisInformation struct {
 	// EstimatedPrintTime is the estimated print time of the file, in seconds.
-	EstimatedPrintTime int `json:"estimatedPrintTime"`
+	EstimatedPrintTime float64 `json:"estimatedPrintTime"`
 	// Filament estimated usage of filament
 	Filament struct {
 		// Length estimated of filament used, in mm
